Extracted settings components into separate views
